fix: only attempt ollama pull if model is not already local
This also refactors the pip_install_if_needed() logic out of the run module and into the pull module.
