+++ pdl-live-react/src-tauri/src/interpreter/pull.rs
@@ -61,8 +61,8 @@ pub async fn pull_if_needed(path: &String) -> Result<(), LoadError> {
     models
         .into_par_iter()
         .try_for_each(|model| match model {
-            m if model.starts_with("ollama/") => ollama_pull(&m[7..]),
-            m if model.starts_with("ollama_chat/") => ollama_pull(&m[12..]),
+            m if model.starts_with("ollama/") => ollama_pull_if_needed(&m[7..]),
+            m if model.starts_with("ollama_chat/") => ollama_pull_if_needed(&m[12..]),
             _ => {
                 eprintln!("Skipping model pull for {}", model);
                 Ok(())
@@ -73,8 +73,21 @@ pub async fn pull_if_needed(path: &String) -> Result<(), LoadError> {
     Ok(())
 }
 
+fn ollama_exists(model: &str) -> bool {
+    match cmd!("ollama", "show", model)
+        .stdout_null()
+        .stderr_null()
+        .run()
+    {
+        Ok(_output) => true,
+        _ => false,
+    }
+}
+
 /// The Ollama implementation of a single model pull
-fn ollama_pull(model: &str) -> Result<(), LoadError> {
-    cmd!("ollama", "pull", model).run().map_err(LoadError::IO)?;
+fn ollama_pull_if_needed(model: &str) -> Result<(), LoadError> {
+    if !ollama_exists(model) {
+        cmd!("ollama", "pull", model).run().map_err(LoadError::IO)?;
+    }
     Ok(())
 }
